backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib
pydantic
python-multipart
alembic
websockets
python-dotenv
cloudinary
pydantic[email]
bcrypt